requirements.txt
# Core dependencies
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
torch>=1.9.0  # Required for PESTO

# Pitch detection
pesto-pitch>=0.1.0

# Audio format support
pydub>=0.25.0
soundfile>=0.10.0  # For additional audio format support

# MusicXML export
musicxml>=0.5.0

# Optional: for audio playback/testing
matplotlib>=3.5.0  # For debugging visualizations

torchvision>=0.10.0  # For CNN
